% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_ent.R
\name{display_ent}
\alias{display_ent}
\title{Visualize entities in text parsed with spacyr.}
\usage{
display_ent(x, deco_ent = "bgYellow", deco_lab = "bgBlue")
}
\arguments{
\item{x}{A spacyr_parsed object}

\item{deco_ent}{A crayon style for entity name, given as a character string}

\item{deco_lab}{A crayon style for entity label, given as a character string}
}
\value{
A character string with styled tokens
}
\description{
Visualize entities in text parsed with spacyr.
}
\details{
x must include the "whitespace_" attribute. Therefore, \code{spacy_parse}
  should be run with the \code{additional_attributes} argument:
  \code{spacy_parse(x, ..., additional_attributes = "whitespace_")}. This
  function is designed to be called for its side effect, which outputs
  a styled character vector to the terminal.
}
\examples{
\dontrun{
txt <- "Larry Page is the co-founder of Google, located in Mountain View, CA."
parsed <- spacyr::spacy_parse(txt, additional_attributes = "whitespace_")
display_ent(parsed)
}
}
